Make HOUSETYPE, ANGLETYPE, CMDTYPE, and DropLocation "enum class".

+++ include/GUI/ObjectInterfaces/MultiUnitInterface.h
@@ -47,7 +47,7 @@ public:
 
 protected:
     MultiUnitInterface() : ObjectInterface() {
-        Uint32 color = SDL2RGB(palette[houseToPaletteIndex[pLocalHouse->getHouseID()]+3]);
+        Uint32 color = SDL2RGB(palette[houseToPaletteIndex[static_cast<int>(pLocalHouse->getHouseID())]+3]);
 
         addWidget(&topBox,Point(0,0),Point(SIDEBARWIDTH - 25,80));
