+++ include/GUI/ObjectInterfaces/UnitInterface.h
@@ -45,7 +45,7 @@ public:
 
 protected:
     explicit UnitInterface(int objectID) : DefaultObjectInterface(objectID) {
-        Uint32 color = SDL2RGB(palette[houseToPaletteIndex[pLocalHouse->getHouseID()]+3]);
+        Uint32 color = SDL2RGB(palette[houseToPaletteIndex[static_cast<int>(pLocalHouse->getHouseID())] + 3]);
 
         mainHBox.addWidget(HSpacer::create(4));
